+++ client/src/components/organisms/study/TodoList.tsx
@@ -1,5 +1,17 @@
+/* eslint-disable no-undef */
 import { FC, memo, Dispatch, SetStateAction } from "react";
-import { List, ListItem } from "@chakra-ui/react";
+import {
+  Tabs,
+  TabList,
+  TabPanels,
+  Tab,
+  TabPanel,
+  Box,
+  Heading,
+  List,
+  ListItem,
+} from "@chakra-ui/react";
+import type { ApolloError } from "@apollo/client";
 import { isWithinInterval, isToday, isBefore, addDays } from "date-fns";
 
 import { TodoWithCheck } from "../../molucules/todos/TodoWithCheck";
@@ -8,23 +20,26 @@ import type { TodoData } from "../../../types/types";
 type Props = {
   todos: Array<TodoData> | undefined;
   loading: boolean;
-  tabType: "全て" | "今日" | "期限切れ";
+  error: ApolloError | undefined;
   onOpen: (e: any) => void;
   setTodoId: Dispatch<SetStateAction<string>>;
 };
+
+// タブのタイプ
+const tabs = ["全て", "今日", "期限切れ"] as const; //as const をつけてreadonlyにする
 
 /**
  * Todoリストを表示するコンポーネント.
  */
 export const TodoList: FC<Props> = memo((props) => {
-  const { todos, loading, tabType, onOpen, setTodoId } = props;
+  const { todos, loading, error, onOpen, setTodoId } = props;
 
   /**
    * Todoをタブのタイプに応じてフィルタリングする.
    *
    * @returns フィルタリングしたtodoの配列
    */
-  const getFilteredTodos = () => {
+  const getFilteredTodos = (tabType: "全て" | "今日" | "期限切れ") => {
     const today = new Date();
 
     if (todos !== undefined) {
@@ -80,19 +95,57 @@ export const TodoList: FC<Props> = memo((props) => {
 
   return (
     <>
-      {loading ? (
-        <>Loading...</>
-      ) : getFilteredTodos()?.length ? (
-        <List>
-          {getFilteredTodos()?.map((todo) => (
-            <ListItem key={todo.id}>
-              <TodoWithCheck {...todo} onOpen={onOpen} setTodoId={setTodoId} />
-            </ListItem>
-          ))}
-        </List>
-      ) : (
-        <p>該当のTodoが存在しません</p>
-      )}
+      <Heading as="h2" size="lg">
+        Todoリスト
+      </Heading>
+      <Box bg="#f5f5f5" padding="5px 24px 10px 24px">
+        <Tabs variant="soft-rounded" isLazy>
+          <TabList>
+            {tabs.map((tab, index) => (
+              <Tab
+                key={index}
+                _focus={{ boxShadow: "none" }}
+                _selected={{ pointerEvents: "none", bg: "green.200" }}
+                _hover={{ backgroundColor: "green.50" }}
+              >
+                {tab}
+              </Tab>
+            ))}
+          </TabList>
+          <TabPanels
+            bg="white"
+            padding="10px 40px"
+            overflow="auto"
+            height="180px"
+          >
+            {tabs.map((tab, index) => {
+              return (
+                <TabPanel key={index}>
+                  {loading ? (
+                    <>Loading...</>
+                  ) : error ? (
+                    <>エラーが発生しました</>
+                  ) : getFilteredTodos(tab)?.length ? (
+                    <List>
+                      {getFilteredTodos(tab)?.map((todo) => (
+                        <ListItem key={todo.id}>
+                          <TodoWithCheck
+                            {...todo}
+                            onOpen={onOpen}
+                            setTodoId={setTodoId}
+                          />
+                        </ListItem>
+                      ))}
+                    </List>
+                  ) : (
+                    <p>該当のTodoが存在しません</p>
+                  )}
+                </TabPanel>
+              );
+            })}
+          </TabPanels>
+        </Tabs>
+      </Box>
     </>
   );
 });
